+++ src/app/shared/interfaces/user.interface.ts
@@ -9,15 +9,3 @@ export interface IUser {
   profilePicture?: string;
   email: string;
 }
-
-export type BaseUserInput = {
-  email: string;
-  password: string;
-};
-
-export type ExtendedUserInput = BaseUserInput & {
-  name: string;
-  lastName: string;
-  telephoneNumber?: string;
-  address?: string;
-};
