+++ src/app/shared/types/user.type.ts
@@ -1,0 +1,11 @@
+export type BaseUserInput = {
+  email: string;
+  password: string;
+};
+
+export type ExtendedUserInput = BaseUserInput & {
+  name: string;
+  lastName: string;
+  telephoneNumber?: string;
+  address?: string;
+};
